flatten root test subtests to two levels
