+++ parser_limits_test.go
@@ -297,73 +297,55 @@ func TestMaxNodeContentSize(t *testing.T) {
 	t.Run("character data", func(t *testing.T) {
 		kinds := []string{"cdata", "comment", "pi", "chardata"}
 
-		t.Run("oversized run fails with ErrNodeContentTooLarge", func(t *testing.T) {
-			t.Parallel()
-			for _, kind := range kinds {
-				t.Run(kind, func(t *testing.T) {
-					t.Parallel()
-					doc := nodeContentDoc(kind, 200)
-					_, err := helium.NewParser().
-						MaxNodeContentSize(64).
-						Parse(t.Context(), []byte(doc))
-					require.Error(t, err)
-					require.ErrorIs(t, err, helium.ErrNodeContentTooLarge)
-				})
-			}
-		})
+		for _, kind := range kinds {
+			t.Run("an oversized run fails with ErrNodeContentTooLarge ("+kind+")", func(t *testing.T) {
+				t.Parallel()
+				doc := nodeContentDoc(kind, 200)
+				_, err := helium.NewParser().
+					MaxNodeContentSize(64).
+					Parse(t.Context(), []byte(doc))
+				require.Error(t, err)
+				require.ErrorIs(t, err, helium.ErrNodeContentTooLarge)
+			})
 
-		t.Run("within-cap run parses fine", func(t *testing.T) {
-			t.Parallel()
-			for _, kind := range kinds {
-				t.Run(kind, func(t *testing.T) {
-					t.Parallel()
-					doc := nodeContentDoc(kind, 32)
-					d, err := helium.NewParser().
-						MaxNodeContentSize(64).
-						Parse(t.Context(), []byte(doc))
-					require.NoError(t, err)
-					require.NotNil(t, d)
-				})
-			}
-		})
+			t.Run("a within-cap run parses fine ("+kind+")", func(t *testing.T) {
+				t.Parallel()
+				doc := nodeContentDoc(kind, 32)
+				d, err := helium.NewParser().
+					MaxNodeContentSize(64).
+					Parse(t.Context(), []byte(doc))
+				require.NoError(t, err)
+				require.NotNil(t, d)
+			})
 
-		t.Run("negative limit disables the cap", func(t *testing.T) {
-			t.Parallel()
-			for _, kind := range kinds {
-				t.Run(kind, func(t *testing.T) {
-					t.Parallel()
-					// A run far past the 10 MiB default still parses when the cap
-					// is disabled with a negative value.
-					doc := nodeContentDoc(kind, 12<<20)
-					d, err := helium.NewParser().
-						MaxNodeContentSize(-1).
-						Parse(t.Context(), []byte(doc))
-					require.NoError(t, err)
-					require.NotNil(t, d)
-				})
-			}
-		})
+			t.Run("a negative limit disables the cap ("+kind+")", func(t *testing.T) {
+				t.Parallel()
+				// A run far past the 10 MiB default still parses when the cap
+				// is disabled with a negative value.
+				doc := nodeContentDoc(kind, 12<<20)
+				d, err := helium.NewParser().
+					MaxNodeContentSize(-1).
+					Parse(t.Context(), []byte(doc))
+				require.NoError(t, err)
+				require.NotNil(t, d)
+			})
 
-		t.Run("a raised cap admits a larger run", func(t *testing.T) {
-			t.Parallel()
-			for _, kind := range kinds {
-				t.Run(kind, func(t *testing.T) {
-					t.Parallel()
-					doc := nodeContentDoc(kind, 4096)
-					// Fails under a small cap...
-					_, err := helium.NewParser().
-						MaxNodeContentSize(1024).
-						Parse(t.Context(), []byte(doc))
-					require.ErrorIs(t, err, helium.ErrNodeContentTooLarge)
-					// ...and parses under a cap large enough for it.
-					d, err := helium.NewParser().
-						MaxNodeContentSize(8192).
-						Parse(t.Context(), []byte(doc))
-					require.NoError(t, err)
-					require.NotNil(t, d)
-				})
-			}
-		})
+			t.Run("a raised cap admits a larger run ("+kind+")", func(t *testing.T) {
+				t.Parallel()
+				doc := nodeContentDoc(kind, 4096)
+				// Fails under a small cap...
+				_, err := helium.NewParser().
+					MaxNodeContentSize(1024).
+					Parse(t.Context(), []byte(doc))
+				require.ErrorIs(t, err, helium.ErrNodeContentTooLarge)
+				// ...and parses under a cap large enough for it.
+				d, err := helium.NewParser().
+					MaxNodeContentSize(8192).
+					Parse(t.Context(), []byte(doc))
+				require.NoError(t, err)
+				require.NotNil(t, d)
+			})
+		}
 
 		t.Run("secure default rejects a run over 10 MiB", func(t *testing.T) {
 			t.Parallel()
@@ -419,32 +401,24 @@ func TestMaxNodeContentSize(t *testing.T) {
 			},
 		}
 
-		t.Run("oversized attribute value fails with ErrNodeContentTooLarge", func(t *testing.T) {
-			t.Parallel()
-			for kind, mk := range bodies {
-				t.Run(kind, func(t *testing.T) {
-					t.Parallel()
-					_, err := helium.NewParser().
-						MaxNodeContentSize(64).
-						Parse(t.Context(), []byte(mk(200)))
-					require.ErrorIs(t, err, helium.ErrNodeContentTooLarge)
-				})
-			}
-		})
+		for kind, mk := range bodies {
+			t.Run("an oversized attribute value fails with ErrNodeContentTooLarge ("+kind+")", func(t *testing.T) {
+				t.Parallel()
+				_, err := helium.NewParser().
+					MaxNodeContentSize(64).
+					Parse(t.Context(), []byte(mk(200)))
+				require.ErrorIs(t, err, helium.ErrNodeContentTooLarge)
+			})
 
-		t.Run("within-cap attribute value parses fine", func(t *testing.T) {
-			t.Parallel()
-			for kind, mk := range bodies {
-				t.Run(kind, func(t *testing.T) {
-					t.Parallel()
-					d, err := helium.NewParser().
-						MaxNodeContentSize(64).
-						Parse(t.Context(), []byte(mk(32)))
-					require.NoError(t, err)
-					require.NotNil(t, d)
-				})
-			}
-		})
+			t.Run("a within-cap attribute value parses fine ("+kind+")", func(t *testing.T) {
+				t.Parallel()
+				d, err := helium.NewParser().
+					MaxNodeContentSize(64).
+					Parse(t.Context(), []byte(mk(32)))
+				require.NoError(t, err)
+				require.NotNil(t, d)
+			})
+		}
 
 		t.Run("cap boundary is strict-greater for the fast path", func(t *testing.T) {
 			t.Parallel()
